LayoutTests/inspector/dom-debugger/event-listener-breakpoints-expected.txt

Tests for Event Listener breakpoints.


== Running test suite: DOMDebugger.Event.Listener
-- Running test case: DOMDebugger.Event.Listener.AddBreakpoint "click"
Creating "click" Event Breakpoint...
Adding "listener:click" Event Breakpoint...
Firing "click" on body...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_click
1: [F] trigger_click
2: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.AddDisabledBreakpoint "click"
Creating "click" Event Breakpoint...
Adding "listener:click" Event Breakpoint...
Disabling "listener:click" Event Breakpoint...
Firing "click" on body...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.RemoveBreakpoint "click"
Creating "click" Event Breakpoint...
Adding "listener:click" Event Breakpoint...
Removing "listener:click" Event Breakpoint...
Firing "click" on body...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.RemoveDisabledBreakpoint "click"
Creating "click" Event Breakpoint...
Adding "listener:click" Event Breakpoint...
Disabling "listener:click" Event Breakpoint...
Removing "listener:click" Event Breakpoint...
Firing "click" on body...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.AddBreakpoint "custom"
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...
Firing "custom" on body...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_custom
1: [F] trigger_custom
2: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.AddDisabledBreakpoint "custom"
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...
Disabling "listener:custom" Event Breakpoint...
Firing "custom" on body...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.RemoveBreakpoint "custom"
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...
Removing "listener:custom" Event Breakpoint...
Firing "custom" on body...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.RemoveDisabledBreakpoint "custom"
Creating "custom" Event Breakpoint...
Adding "listener:custom" Event Breakpoint...
Disabling "listener:custom" Event Breakpoint...
Removing "listener:custom" Event Breakpoint...
Firing "custom" on body...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.AllEventsBreakpoint
Adding "listener" Event Breakpoint...
Firing "click" on body...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_click
1: [F] trigger_click
2: [P] Global Code
Firing "custom" on body...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_custom
1: [F] trigger_custom
2: [P] Global Code
PASS: Should have paused twice.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.AddMultipleBreakpoints
Creating "click" Event Breakpoint...
Adding "listener:click" Event Breakpoint...
Firing "click" on div#x...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleX_click
1: [F] clickX
2: [P] Global Code
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleBody_click
1: [F] clickX
2: [P] Global Code
PASS: Should have paused twice.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.Listener.DisableBetweenDispatches
Creating "click" Event Breakpoint...
Adding "listener:click" Event Breakpoint...
Firing "click" on div#x...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleX_click
1: [F] clickX
2: [P] Global Code
Disabling breakpoints...
PASS: Should have paused once.
-- Running test teardown.
